# SubastasWeb/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Subastas",
        "description": "Documentación de la API de Subastas",
        "contact": {
            "email": "contacto@subastas.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/casa-remates": {
            "get": {
                "tags": [
                    "CasaRemates"
                ],
                "summary": "Obtener todas las casas de remate",
                "operationId": "a4047afdf0b75fad619b63b3d348c406",
                "responses": {
                    "200": {
                        "description": "Lista de casas de remate"
                    }
                }
            },
            "post": {
                "tags": [
                    "CasaRemates"
                ],
                "summary": "Crear una nueva casa de remate",
                "operationId": "52d5f60b97188f7b0cf41bc1e3e7fd2e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "idFiscal"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer"
                                    },
                                    "idFiscal": {
                                        "type": "string",
                                        "maxLength": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Casa de remate creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/casa-remates/{usuario_id}": {
            "get": {
                "tags": [
                    "CasaRemates"
                ],
                "summary": "Obtener una casa de remate por ID",
                "operationId": "ab33a9d395cc88238e40f2514fb00609",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Casa de remate encontrada"
                    },
                    "404": {
                        "description": "Casa de remate no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "CasaRemates"
                ],
                "summary": "Actualizar una casa de remate",
                "operationId": "d5a1a2c11557fb0fa4eb8d44a7c283eb",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idFiscal": {
                                        "type": "string",
                                        "maxLength": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Casa de remate actualizada"
                    },
                    "404": {
                        "description": "Casa de remate no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CasaRemates"
                ],
                "summary": "Eliminar una casa de remate por ID",
                "operationId": "9f13e312194c7364ddcc9d387ffe94a6",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Casa de remate eliminada"
                    },
                    "404": {
                        "description": "Casa de remate no encontrada"
                    }
                }
            }
        },
        "/api/casa-remates/{usuario_id}/asociar-rematadores": {
            "post": {
                "tags": [
                    "CasaRemates"
                ],
                "summary": "Asociar rematadores a una casa de remate",
                "operationId": "f682598b87693062828a9a121300df5b",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rematadores"
                                ],
                                "properties": {
                                    "rematadores": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rematadores asociados exitosamente"
                    },
                    "404": {
                        "description": "Casa de remate no encontrada"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/api/casa-remates/{usuario_id}/calificar": {
            "post": {
                "tags": [
                    "CasaRemates"
                ],
                "summary": "Agregar una calificación a una casa de remate",
                "operationId": "b2b166a84b8e6aea6119d08118644e2d",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "description": "ID de la casa de remate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valor"
                                ],
                                "properties": {
                                    "valor": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calificación agregada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string"
                                        },
                                        "promedio": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "total_opiniones": {
                                            "type": "integer"
                                        },
                                        "estrellas": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Casa de remate no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}